Move dbs-info service to libs
