+++ src/commands/db/ls.ts
@@ -1,7 +1,7 @@
 import { Command } from '@effect/cli';
 import { Array, Console, Effect } from 'effect';
 import { initializeUrl } from '../../index';
-import { CouchDbInfo, getAllDbsInfo } from '../../services/couch/dbs-info';
+import { CouchDbInfo, getAllDbsInfo } from '../../libs/couch/dbs-info';
 import { getDisplayDictByPid } from '../../services/couch/active-tasks';
 
 const getDbDisplay = ({ info: { db_name, doc_count } }: CouchDbInfo) => ({
