+++ test/services/monitor.spec.ts
@@ -4,8 +4,8 @@ import { expect } from 'chai';
 import { MonitorService } from '../../src/services/monitor';
 import { CouchNodeSystem, CouchNodeSystemService } from '../../src/services/couch/node-system';
 import sinon, { SinonStub } from 'sinon';
-import { CouchDbInfo } from '../../src/services/couch/dbs-info';
-import * as CouchDbsInfo from '../../src/services/couch/dbs-info';
+import { CouchDbInfo } from '../../src/libs/couch/dbs-info';
+import * as CouchDbsInfo from '../../src/libs/couch/dbs-info';
 import { CouchDesignInfo, CouchDesignInfoService } from '../../src/services/couch/design-info';
 import { LocalDiskUsageService } from '../../src/services/local-disk-usage';
 import { createDbInfo, createDesignInfo, createNodeSystem } from '../utils/data-models';
